// src/features/pages/client/client.model.ts
export class Client {
    id: number;
    code: string;
    type: string;
    name: string;
    country: string;
    countryName: string;
    payment_condition: string;
    payment_condition_value: string;
    firstname: string;
    lastname: string;
    reference: string;
    registry_number: string;
    vat_number: string;
    address: string;
    zipcode: string;
    city: string;
    mail: string;
    phone: string;
    ebill_type: string;
    ebill_identifier: string;
}